Move util code into xdp-trafficgen
